Update puppet package install
There is something wrong with package dependencies
in the packages from PL.

Specifically, that it tries to install the latest
version of puppet-common regardless of what version
we specify for Puppet.

This leads to package version conflicts when we
want to specify the exact package version to install.

This commit ensures that puppet-common is installed as the
correct version before puppet, which seems to get around the issue.