Remove cobbler from site_manifest
This commit removes the cobbler nodes from
the site manifest. They should not always be
added to the node, they should only be added
when cobbler is being managed as a part of
the build node.

Additionally, you should be able to dynamically
create nodes as a part of the cobbler node's
hiera data.