remove references to coi test_file
This commit is a step towards deprecating
the coi module.

Updates role to replace the coi test file
class with the ones that it was composed of.

Updates relevent data mapping and hiera configs.

Updates Puppetfile to use the latest version of
puppet-coi that removes the deprecated classes.

removes usage of the build test class that was
using these now deprecated classes.