Fix dep issue on redhat
I previously made a modification to the puppet packages
in order to resolve from missing package dependencies
for the Ubuntu packages.

These changes created a dependency loop on the puppet package.

This commit refactors the way in which the puppet packages are
installed so that htey should work on both platforms.