Add fragments to 2_role node definitions.
These are used to compose the script that will be
passed to cloud-init for each node. Note that there
is a primitive signalling system to ensure things
happen in the correct order.

The script that will be generated for a particular host
under a particular scenario can be viewed by using
stack-builder's frag command like so:

sb frag build-server -s 2_role