Support data mappings of runtime ip addresses
Add support for specifying hiera keys that
should have the value of a given interface of a
given node, for example the cobbler_node_ip
needed by hiera and build_server_ip as needed
by setup.pp. These values will be in both config
and user yaml on all nodes.