Add puppet_run_type config to setup.pp
Currently, setup.pp configures every node as
if they were a master (or a node that needs to
run puppet apply). This configuration is unnecessary
and even undesirable for agents b/c it deploys
hiera data they don't need and even enables then to use
a node terminus.

This patch adds a variable to setup.pp called puppet_run_type
that can be used to determine if the node is being
pre-confgigured as a agent or a master.

It also updates the Vagrantfile as well as the bare-metal install
scripts.