Add rebuild command
The rebuild command can be used for debugging, as it will build
only the nodes that do not exist for a test run. The most common
use would be to keep the build node while terminating other
instances, then using rebuild to recreate them. Be careful of
puppet certificate errors if you do this - cert cleaning on the
master may be required.

Also added a check to make sure the right CI network is chosen,
which requires the keystone module in order to confirm which tenant
we are running as.