do not include white space in ceph user name
It will break with:

ProcessExecutionError: Unexpected error while running command.
Command: rbd create --pool rbd --size 1024 volume-32b2980d-0060-42a5-98fd-7408a03a8570 --new-format
Exit code: 1
Stdout: ''
Stderr: 'rbd: extraneous parameter cephx\n'

Which is confusing because the cephx parameter does not show in the
command displayed. It actually comes from the environment because
/etc/init/cinder-volume.override has

    env CEPH_ARGS="--id no user"

instead of

    env CEPH_ARGS="--id 'no user'"

Although an option would be to make sure it is quoted everywhere,
it is likely to break in so many places that it's probably not worth the
effort anyway.

Signed-off-by: Loic Dachary <loic@dachary.org>