Add Proper Rabbit HA Support for Havana
Previously, rabbit_hosts was only being specified for Quantum and
not Neutron.  THis caused the rabbit_hosts parameter not to be
configured when using Havana.  Additionally, rabbit cluster
parameters were being exposed directly from the rabbitmq module.
This is unneeded and not optimal, since these parameters are
conditionals within the nova module.