Remove stack-builder and add all_in_one scenario
Stack-builder tools have been broken out into their own repository
and can now be downloaded via pip. This patch updates the README
and removes the stack-builder tools, as well as adding support
for the all_in_one scenario

Scenariobuilder 0.1.6 moves from fragments to jinja2
templating, and the format of specifying node
order has changed as a result. This also adds a list
of the config options available to configure the
bootstrap script in config.yaml

Change-Id: I0b5f5ddc052b71bb0871e63ff01d1a3638a5d244