Add heat to modules.list
A previous commit added support for Heat to
puppet_openstack_builder.  However, the change only works when
using install_type=git because the heat wasn't added to
modules.list (which is read when determining what packages need
to be installed when using install_type=deb).  This commit adds
heat to the modules.list file so that the puppet-heat package
gets installed when using install_type=deb.

Change-Id: I5cec210f6da9d4539230595d57b281f0acf56028