Clean up install scripts
setup.sh script now should only configure the node with
necessary dependencies and hostname info. The extra
cruft to check out puppet_openstack_builder repo is
removed.

install.sh now only applies the configuration if the
scenario is all-in-one. For the rest, user config
update is required. So will exit with a message
suggesting to do the same and run puppet apply.

Change-Id: If2ff86ead554379dbc332a256f162d1219090a1a