Add graphite role to build node
Previously, collectd was set up on many node scenarios and was
being configured to send data to a carbon instance on the build
node.  However, the build node wasn't actually running graphite,
which rsulted in a good bit of logspam and obviously didn't
help anyone collect any data.  This commit adds graphite to the
build node role so that something is listening for incoming data
from collectd.

Change-Id: I6a5f5c0c11bd9f3a61ec7f20d0b35cddf83b50c6