Added contrib directory for related tools.
I think it makes sense to have a contrib directory for tools that are related
to this installation. I've added a storage sub-directory that contains a
keygen tool for generating ceph keys for initial installation. Right now,
the ceph keys are hard-coded in user.common.yaml. Unique keys should be generated
for each installation, rather than using the same set of keys.

Change-Id: Id44fd05387f3c3caba5fef3dc23aadc5138f9aa0