Expose time zone in user.common.yaml
The puppet-coi module has a parameter for setting the time zone
as well as a list of NTP servers, but currently user.common.yaml
only has data for the servers.  This commit adds a variable for
the time zone so that users know they can modify it according
to what time zone they're in.  It defaults to UTC as that's
also the default for lower-level classes.

Requires a puppet-coi containing 0ccb122 to be useful.

Change-Id: I6914bb86d916e262de532e42227b978f4a8eca1c