Add heat templates for 2_role scenario
This commit adds some heat templates that can be used
in conjunction with scenariobuilder to create stacks that
deploy puppet_openstack_builder in VMs on an openstack
cloud.

The template and bootstrap cannot be used on their own,
as user and global config is must be inserted into
the metadata fields of each node informing them of the
desired configuration.

Two sample config files are provided under contrib to
help users get started.

Change-Id: I1b9365a8d6e5afc96e720836aadc07df5595d965