Set vncserver_proxyclient_address to real IP address
The install.sh script sets the vncserver_proxyclient_address in
nova.conf to 0.0.0.0 for AIO installs.  That may be ok for a true
AIO, but we often see people use AIO as a combined control/compute
node and then add additional compute nodes.  In that case, consoles
become unusable due to the lack of an actual address.  This patch
sets the address to that IP on the default interface.

Change-Id: I3c12e1732797eb7953eefda14f6557c5d976f863
Closes-Bug: 1269241