Add -y to apt-get command
The apt-get command that installs puppet modules when using a package-
based install lacks the -y flag.  On most systems this doesn't
seem to be a problem since dependencies are already installed.
However on some systems this can cause the install to fail out
if a dependency is needed (thus cause apt-get to interactively
ask the user if it's ok to install the dependency as well).

Thanks to Jon Skarpeteig for reporting the issue.

Change-Id: I9dddde04cf7f8ef2e777ff9f1bc9feed1cfd6c83