Don't map controller_hostname to an IP address
The current data mappings set coe::base::controller_hostname
to an IP address.  This results in an invalid entry in
/etc/hosts containing two IP addresses rather than an IP address
and a hostname.  The problem doesn't affect the full_ha
scenario as it overrides the setting in user.full_ha.yaml.  This
patch removes the broken mapping and adds a similar config to
user.common.yaml, resulting in a clean /etc/hosts file.

Change-Id: Ibdb436ce0f5d08fd6acbadebc7323ac374a46207
Partial-Bug: #1274263