Don't use FQDN in build_node_name
The build_node_name setting in two of the same yaml files
is fully qualified--but shouldn't be, as using an FQDN may
result in entries in /etc/hosts that have the domain name
twice:

hostname.domain.com.domain.com

This patch modifies the sample build_node_name settings in
the yaml files in which the values are currently FQDNs.
Additionally, it modifies the install scripts to ensure
that it gets the short name when running an all_in_one
buildout.

Change-Id: I9b43a467afdfbcedf4ff9d91f60f4597aa107c77
Closes-Bug: #1276260