Fix backward /etc/hosts entry
The setup.pp manifest populates an entry into /etc/hosts, but
it puts the short hostname before the canonical hostname.
According to "man hosts" and other documentation, this is incorrect:
the canonical hostname should come first.  This patch corrects
the order and corrects style violations in the host entry.

Change-Id: If457b8eb74fb645d66d35d38bc2e06b683b9f9b3
Partial-Bug: #1274263