Remove references to ftpeng.cisco.com
The ftpeng hosting site is being replaced by Cisco IT, and new files
cannot currently be uploaded to it. While previous releases of COI can
continue to use ftpeng.cisco.com as a package mirror, new releases of
COI should not contain any references to it so the site won't contain
any updates from after Feb 14th. Alternate FTP-based hosting is being
looked into, but isn't likely to be ready before the h.2 release. Just
to note: starting in at least h.0 the default package hosting site in
COI is http://openstack-repo.cisco.com and demand for FTP hosting is
very small anyway, so it's unlikely that this will affect anyone.

Change-Id: I7ffbed471473fe24f302d6d2e69d3e9d55c00c72
Partial-Bug: #1281784