Use the short hostname for the AIO role mapping
In the AIO scenario the build_server is set to the short hostname,
while the role mapping for the AIO uses the full name. This causes
a mismatch which leads to the AIO scenario no-oping on the first run.
This change ensures that the hostname matches exactly what is in the
role mappings file.

Change-Id: Ia3fd5ef716695ff80cea4187a488b2705e518d3e