Remove deprecation warning/unify package_ensure
Puppet-coi calls the ntp class with a parameter that has been
deprecated in newer versions of the PuppetLabs NTP module that
manages whether the NTP package gets updated as new packages are
made available, or is merely installed once.  Presumably deployers
will want to manage that setting the same way they manage the
settings for other packages in the system, so it makes far more
sense to map it to the package_ensure variable at the composition
layer.  This patch does that.  Note also that a separate patch
to puppet-coi has removed the deprecated param.

Change-Id: I6d0eae7d6fead8ccd64bfa7fae1ec5694580f627
Partial-Bug: #1281896