Fix bad param to setup_test_volume class
When setting up the iSCSI backend, we also call the
setup_test_volume class from puppet-cinder.  It takes a param
called 'volume_name', but we're trying to send it 'volume_group'.
The bug has no negative effect since the value we set matches
the default value for volume_name anyway, but users that read
the YAML will probably be confused if they try to change the
volume name and don't see it take effect.

Change-Id: Iabb6d2bd781be13d3caecc63b18d2649778c3460
Closes-Bug: #1284319