Expose libvirt disk_cachemodes setting
A recent change to the puppet-nova module allows users to specify
an array of settings to propagate into the disk_cachemodes setting
of nova.conf.  This change exposes the new capability to end users
so that those who need to tune IO settings can easily do so.  The
patch leaves the new setting commented out by default since this
will maintain the current behavior of puppet_openstack_builder with
regards to the disk_cachemodes setting in nova.conf (and since the
defaults are probably what most users will want).

Change-Id: I514182cbd1f107c9264c037ca6e3431b1e04bcc5
Implements: blueprint expose-disk-cachemodes