Remove controller and compute class groups.
These two class groups are composed from other class
groups. The forthcoming rewrite of the data model to
support pure hiera does not support this, and since
controller and compute are basically just restating
the role names and hiding what is actually on the node,
this patch removes them and places the set of groups
they were abstracting in their place.

Change-Id: Ice72703cbd5523e03876d09513e42c2c0ad050d7