Removed cinder and glance scenario dirs
These dirs contained files that would allow a user to add additional
classes to a specific scenario. This causes a problem when dealing with
scenarios like full_ha, where the services are broken out across multiple
nodes. The classes and class_group declarations are already handled in
the scenario files, they are not needed here.

Change-Id: I29c1477edbb7967a4a5c1e15d12572a193bdae94