Explicitly disable Neutron DB sync by default
Bug #1288975 reports that Neturon database syncs are enabled
by default, but may cause problems in many cases where plugins
set database connection information in neutron.conf rather than
in the plugin's ini file.  This is known to be the scenario for
the OVS plugin, which is the default plugin for
puppet_openstack_builder.  Running the DB sync is also known
to not be necessary.  This patch adds a knob allowing users to
specify whether or not to run the sync, and sets it to false by
default.

Change-Id: I84821022743f3a80f253d8d823be5383f5df5194
Related-Bug: #1288975