Fix outdated openstack_repo_location for HA
Some time back the expected value of the openstack_repo_url
parameter was changed to make it more flexible.  The update was
done in such a way that the puppet-coe module would detect if the
old format was used and correct it on the fly, emitting a warning
in the process to let the user know what it did.  A composition
layer change was made to set the new format in
ebfc0a2801f63982ecc83883f8f1885da5692fd4, but only in
data/hiera_data/vendor/cisco_coi_common.yaml.  However, the value
is also explicitly set for HA scenarios in
data/hiera_data/enable_ha/true.yaml and that file was never
updated with the new format.  Thus, users of HA scenarios see
a warning emitted.  This patch updates the setting for HA
scenarios so the warning goes away.

Change-Id: I13bc10ea4ccbc064e8d76468b6bdae9df599f2d8
Closes-Bug: #1292152