Add support for puppet-mysql 2.2
The PuppetLabs MySQL module has been updated to a much cleaner
version, but is not completely backward compatible with the old
0.9 version we're presently using.  Changes have been made to
puppet-nova, puppet-neutron, puppet-heat, etc to add a parameter
allowing users to select which of the two supported versions of
puppet-mysql to use.  This patch adds a data mapping that allows
users to choose which version they want to use via a single variable.
It defaults to 2.2 as we intend to use that going forward.

Change-Id: I6cdcde7d1f67acf7570f2ef15c0997b67f0f144c