Fix potential duplicate resource error in AIO
When cobbler is enabled in the AIO scenario, there is a potential
duplicate resource conflict when defining hosts in the /etc/hosts
file.  Puppet-coi and puppet-coe both create host definitions
(one for the control node and one for the build node...which for
AIO are actually the same).  Install.sh creates a user.yaml when
called with the all_in_one scenario set and populates the control
node addresses into it...but doesn't populate cobbler_node_ip.
This can lead to a duplicate resource error because although the
two modules both use ensure_resource, the IP addresses aren't set
consistently in user.yaml for the AIO scenario.  This patch corrects
the problem by having install.sh set cobbler_node_ip to the same
IP address used for the controller.

Change-Id: Iafbe9b2603ab782d1ab8fd666312ca14bb6dd7f0
Closes-Bug: #1308257