Configure MySQL before db sync runs
When using puppet-mysql 2.2, the restart and bind-address options need to
be passed so that MySQL server will be configured properly before db sync
runs are made for the various OpenStack services.

Also include max-connections, max_connect_errors, and key_buffer_size as
generally useful parameters that together provide an example to users of
how to supply MySQL config options through the composition layer.

Change-Id: I0b607a56b63f51bc698c23b1797d81f9d7dca506
Closes-Bug: #1304579