Fix region data mappings
The region data mappings contain a couple of typos (single ':'
instead of two separating class names), and a few more need to be
added for classes that have added region params over time.  This
patch corrects errant mappings and adds missing ones.

Thanks to Ken Schroeder for the find and suggested fix.

Change-Id: I5fc5ae1b41ddc3fc5ddf5d7a481a5191a590509e
Closes-Bug: #1308366