Map glance_api_servers for cinder
Puppet_openstack_builder has a mapping for the
nova::glance_api_servers parameter, which tells nova where to
find the glance API service.  Cinder takes a similar parameter,
but we don't currently have a mapping for it (which can result
in errors when creating a volume using an image as the source).
This patch adds a mapping so that Cinder knows where to find
the glance API.

Change-Id: If168fdeddfb06f63186c2ad4759f1656c6055905
Closes-Bug: #1284772