Add dhcp firewall rule to controller
There is a new class in openstacklib to add the forwarding
rule needed to run the linuxbridge dhcp agent with a default
drop firewall. This patch adds that class to the control_firewall
class group. Note that the new class does nothing by default, and
can be safely included even when not using linuxbridge.

Change-Id: Ic2d2c749734358da9e96cd18fcf353005fa47b38