Add support for nova-neutron interaction
In Icehouse, a new notification system was introduced allowing
Neutron to communicate with Nova about port status and data.
Configuring this feature requires some extra setup, including
querying keystone for the UUID of the nova admin tenant and
placing it in the neutron.conf file.  Since the new interaction
system is on by default, failing to configure it results in
fatal errors when attempting to launch an instance.  This patch
adds the necessary composition layer changes needed to support
the new notification system.  It also fixes a bug in which the
admin_tenant and service_tenant parameters weren't actually being
mapped to anything (these are also needed to configure the new
interaction system) and sets them to the default values that
ended up getting deployed because of the bug.

Change-Id: I5157b59ab1348bbe0854b6e88a0f60affdd6600a
Closes-Bug: #1302163