Minor fixups for the stacktira bootstrap script
- ruby-augeas requires gcc since it's a native extension.
It was also previously bundled in with the puppet install, which made
its state invisible to the installer. Before installing puppet,
this patch installs gcc if needed, installs the gem, then removes gcc

- epel was being handled in a very blunt manner for libyaml install.
This patch makes a quick search to see if epel is already enabled on
the node, and installs it only if it isn't there.

Change-Id: Ib67fdd86089855cead228725d7cbc2e61634ff53