Fix Cinder deprecation warning
The puppet-cinder module has deprecated the sql_connection
parameter and replaced it with a parameter called
database_connection to be more consistent with other modules.
The puppet_openstack_builder data mappings haven't been updated,
so users see a deprecation warning.  This patch updates the
data mappings to get rid of the warning.

Change-Id: I947eb7fa7718a720a72b7275af0aad0376d41776
Closes-Bug: #1317160