Fix iscsi_ip_address to work automatically for AIO
In https://review.openstack.org/#/c/101589/ a change was made
to allow compute nodes to connect to iscsi volumes that don't
reside on the same physical host. However, the default value used
is the IP address on eth1, which causes a minor headache for many
people deploying AIO with a single NIC (usually eth0).  This patch
adds a line to the install.sh script that sets a value in user.yaml
that works for such cases.

Change-Id: I6d968ec5fb9bea56159c48e754e94e53ab0b1bb8
Closes-Bug: #1344074