Add horizon mappings
Horizon will by default use the fqdn of the controller
node. This is not correct in the case where the fqdn
being used is from a non-public interface, or for the
case where a virtual IP sits in front of the dashboard
and has its own name.

This patch uses controller_public_address for the horizon
fqdn and servername.

Change-Id: I75a00c0f1b9a265a8c103109971b018def903196